Wordmatch
I was given a puzzle. The puzzle's clue was a letter; the puzzle itself was a
series of numbers and punctuation marks. The trick was to reverse the clue and
then each letter in the puzzle was the index of the clue word that went there
from the plaintext. Not the cleanest, but it'll do.